views/user: Merge into "users"
